chatbot-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
python-dotenv==1.0.1
qdrant-client==1.12.0
langchain==0.3.7
langchain-openai==0.2.0
langchain-openrouter==0.1.3
langchain-cohere==0.3.0
langchain-community==0.3.7
openai==1.55.0
cohere>=4.21
openrouter-python==1.0.0
transformers==4.45.0
torch==2.5.0
sentence-transformers==3.1.0
